requirements.txt
requests
pillow
reportlab
python-docx
